+++ example/lib/example_inputtext.dart
@@ -14,7 +14,7 @@ class ExampleInputText extends StatelessWidget {
             size: BsInputSize.sm,
             hintText: 'Small input',
             controller: TextEditingController(),
-            validators: [BsValidator.required],
+            validators: [BsInputValidator.required],
           ),
         ),
         BsCol(
@@ -24,7 +24,7 @@ class ExampleInputText extends StatelessWidget {
             size: BsInputSize.md,
             hintTextLabel: 'Medium input',
             controller: TextEditingController(),
-            validators: [BsValidator.required],
+            validators: [BsInputValidator.required],
           ),
         ),
         BsCol(
@@ -66,7 +66,7 @@ class ExampleInputText extends StatelessWidget {
             hintTextLabel: 'Large input',
             maxLines: 5,
             controller: TextEditingController(),
-            validators: [BsValidator.required],
+            validators: [BsInputValidator.required],
           ),
         ),
       ],