+++ example/lib/main.dart
@@ -2,9 +2,11 @@ import 'package:bs_flutter/bs_flutter.dart';
 import 'package:bs_flutter_example/example_alert.dart';
 import 'package:bs_flutter_example/example_badge.dart';
 import 'package:bs_flutter_example/example_buttons.dart';
+import 'package:bs_flutter_example/example_datatables.dart';
 import 'package:bs_flutter_example/example_grid.dart';
 import 'package:bs_flutter_example/example_inputtext.dart';
 import 'package:bs_flutter_example/example_modal.dart';
+import 'package:bs_flutter_example/example_selectbox.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -45,6 +47,8 @@ class _MyAppState extends State<MyApp> {
               ExampleInputText(),
               ExampleAlert(),
               ExampleModal(),
+              ExampleSelectBox(),
+              ExampleDatatables(),
             ],
           ),
         )),