+++ lib/src/others/bs_container.dart
@@ -1,3 +1,4 @@
+import 'package:bs_flutter/bs_flutter.dart';
 import 'package:bs_flutter_utils/bs_flutter_utils.dart';
 import 'package:flutter/cupertino.dart';